64-bit patches submitted by awoodland

+++ src/backgrnd.c
@@ -39,7 +39,7 @@ JE_word megaDataSeg, megaDataOfs, megaData2Seg, megaData2Ofs, megaData3Seg, mega
 JE_word mapX, mapY, mapX2, mapX3, mapY2, mapY3;
 JE_byte **mapYPos, **mapY2Pos, **mapY3Pos;
 JE_word mapXPos, oldMapXOfs, mapXOfs, mapX2Ofs, mapX2Pos, mapX3Pos, oldMapX3Ofs, mapX3Ofs, tempMapXOfs;
-unsigned int mapXbpPos, mapX2bpPos, mapX3bpPos;
+intptr_t mapXbpPos, mapX2bpPos, mapX3bpPos;
 JE_byte map1YDelay, map1YDelayMax, map2YDelay, map2YDelayMax;
